Extract ReaderStreamTitleView
